fix: supervisord running detection logic
